Upgraded select npm modules
There were some small changes necessary in the code to work with
the current version of traceur.  One of which was appending '.js'
to module names.  This may be due to some low level change in the
webLoader since we are bootstrapping the TraceurLoader manually.
Hopefully we can put this back soon.